Rename Charging Module services
Now that our Charging Module-specific services live in the `charging-module` folder, having `charging-module` in their filenames is redundant, so we remove it
